Move /my, /language, /birthday, /link commands into /profile, support project and 30d tracking for user activity
